fix: revert build config changes
This reverts part of commit e622fa9d79a3a73b322f291d01ef08ab2f03ba45
which made changes to the build config to work with yarn workspaces.
Those changes broke the Docker build for the server. This commit
leaves the server build as it was, and does not restore client/yarn.lock
which is suboptimal as both client and server are now effectively
building without lockfiles, but brings client to parity with server and
fixes the build for now. We should find an optimal fix that uses lock files
for both client and server and works properly with yarn workspaces.